# Improve speed of the ChallengeSolves API endpoint