* Improve speed of the ChallengeSolves API (`/api/v1/challenges/[challenge_id]/solves`) endpoint

From time testing there is a significant improvement:

Before:

```
690 ms ± 47.5 ms per loop (mean ± std. dev. of 10 runs, 3 loops each)
684 ms ± 45.1 ms per loop (mean ± std. dev. of 10 runs, 3 loops each)
712 ms ± 79 ms per loop (mean ± std. dev. of 10 runs, 3 loops each)
761 ms ± 108 ms per loop (mean ± std. dev. of 10 runs, 3 loops each)
```

After:

```
125 ms ± 25.6 ms per loop (mean ± std. dev. of 10 runs, 3 loops each)
115 ms ± 7.94 ms per loop (mean ± std. dev. of 10 runs, 3 loops each)
127 ms ± 17.6 ms per loop (mean ± std. dev. of 10 runs, 3 loops each)
133 ms ± 27.2 ms per loop (mean ± std. dev. of 10 runs, 3 loops each)
```

We should consider the potential performance gains vs risks of making all relationships use joined loading instead of select lazy loading.